# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemone Diamond Catalog Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface JewelryItem {
  srNo: number;
  title: string;
  weight10k: number;
  weight14k: number;
  weight18k: number;
  centerDiamondWeight: number;
  sideDiamondWeight: number;
  /** Base64 encoded image from the Excel file */
  imageBase64?: string;
  /** MIME type of the image (e.g. image/jpeg) */
  imageMimeType?: string;
}

export interface ParsedExcelData {
  items: JewelryItem[];
  totalRows: number;
}

export interface PricingConfig {
  /** Gold price per gram in INR */
  goldPriceINR: number;
  /** Diamond price per carat in USD */
  diamondPriceUSD: number;
  /** USD to INR conversion rate */
  usdToInrRate: number;
  /** Labour charge per gram of metal */
  labourPerGram: number;
  /** Fixed wastage amount (B2B only) */
  wastageFixed: number;
  /** Handling charge percentage */
  handlingPercent: number;
  /** Profit margin percentage (B2C only) */
  profitPercent: number;
  /** Admin charge percentage (B2B only) */
  adminChargePercent: number;
}

export type GenerateCatalogRequestCatalogType =
  (typeof GenerateCatalogRequestCatalogType)[keyof typeof GenerateCatalogRequestCatalogType];

export const GenerateCatalogRequestCatalogType = {
  B2B: "B2B",
  B2C: "B2C",
} as const;

/**
 * Gold karat for the catalog
 */
export type GenerateCatalogRequestKarat =
  (typeof GenerateCatalogRequestKarat)[keyof typeof GenerateCatalogRequestKarat];

export const GenerateCatalogRequestKarat = {
  "10K": "10K",
  "14K": "14K",
  "18K": "18K",
} as const;

export interface GenerateCatalogRequest {
  items: JewelryItem[];
  pricingConfig: PricingConfig;
  catalogType: GenerateCatalogRequestCatalogType;
  /** Whether to show individual charge breakdown in the catalog */
  showItemizedCharges: boolean;
  /** Gold karat for the catalog */
  karat: GenerateCatalogRequestKarat;
}

export type UploadExcelBody = {
  file: Blob;
};