# lib/api-zod/src/generated/types/generateCatalogRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemone Diamond Catalog Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateCatalogRequestCatalogType } from "./generateCatalogRequestCatalogType";
import type { GenerateCatalogRequestKarat } from "./generateCatalogRequestKarat";
import type { JewelryItem } from "./jewelryItem";
import type { PricingConfig } from "./pricingConfig";

export interface GenerateCatalogRequest {
  items: JewelryItem[];
  pricingConfig: PricingConfig;
  catalogType: GenerateCatalogRequestCatalogType;
  /** Whether to show individual charge breakdown in the catalog */
  showItemizedCharges: boolean;
  /** Gold karat for the catalog */
  karat: GenerateCatalogRequestKarat;
}