# lib/api-zod/src/generated/types/jewelryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemone Diamond Catalog Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface JewelryItem {
  srNo: number;
  title: string;
  weight10k: number;
  weight14k: number;
  weight18k: number;
  centerDiamondWeight: number;
  sideDiamondWeight: number;
  /** Base64 encoded image from the Excel file */
  imageBase64?: string;
  /** MIME type of the image (e.g. image/jpeg) */
  imageMimeType?: string;
}