# lib/api-zod/src/generated/types/parsedExcelData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemone Diamond Catalog Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { JewelryItem } from "./jewelryItem";

export interface ParsedExcelData {
  items: JewelryItem[];
  totalRows: number;
}