# lib/api-zod/src/generated/types/pricingConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gemone Diamond Catalog Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface PricingConfig {
  /** Gold price per gram in INR */
  goldPriceINR: number;
  /** Diamond price per carat in USD */
  diamondPriceUSD: number;
  /** USD to INR conversion rate */
  usdToInrRate: number;
  /** Labour charge per gram of metal */
  labourPerGram: number;
  /** Fixed wastage amount (B2B only) */
  wastageFixed: number;
  /** Handling charge percentage */
  handlingPercent: number;
  /** Profit margin percentage (B2C only) */
  profitPercent: number;
  /** Admin charge percentage (B2B only) */
  adminChargePercent: number;
}